# Compiler: don't merge union types if there's a type annotation