Fixes #9050

It would be nice to test the ecosystem if this goes through because it's a big change.