@asterite 2020-04-13
> There was a change last time and we reverted it because of one project's really bad compile-time regression
> could we test that project again?

We could, but it's different. The old change made all unions not to unify into their parent. This time it's controlled: only when you specify an instance/class var as a union. Give that the feature doesn't exist yet, nobody is doing that, so the only way to try this is to modify existing projects to use the feature and try it with a compiled compiler. It's not straight-forward.

I tried it on the compiler, by making `TypeDeclaration#@var` to be a union. Here are the times:
```
Before:
63.98s user 4.39s system 204% cpu 33.353 total
After + no changes to code:
64.01s user 4.32s system 191% cpu 35.649 total
After + changes:
65.57s user 4.22s system 203% cpu 34.226 total
```

The times are similar. There are variances but I don't think it's because of the changes but rather what the computer was doing (I'm listening to music, doing other stuff).

I think this shouldn't affect performance.