@bcardiff 2020-04-13
I found it confusing that the semantics will change depending on how things are written and not what values are involved. This can lead to hard bugs in case the expressions are not preserved along the way.

If I would need to tag some times as in this examples I would use modules.

```crystal
module Foo; end

class Base; end

class One < Base
  include Foo
end

class Two < Base
  include Foo
end

class Three < Base; end

m = Two.new.as(Foo)

x = case m
    when One
      1
    when Two
      2
    end
```

But the new exhaustive case does not cover this, forcing a `when Foo` or `else`. If this is handled, what would be the downsides of this approach?