@asterite 2020-04-13
> But the new exhaustive case does not cover this, forcing a when Foo or else. If this is handled, what would be the downsides of this approach?

The downside is that if this is library code, and someone uses your library and adds a new type `Foo`, the code will stop compiling. And given that this is in a library code, there's no way to fix it.

Exhaustiveness check can only be done if the types are guaranteed to be a determined set. With type hierarchy or included modules, this is not the case.

In Ruby's sorbet there's the concept of [sealed](https://sorbet.org/docs/sealed), which means only types defined directly in that module right there are the only possible types, and these can't be extended. In Crystal we could do something similar:

```crystal
@[Sealed]
module Foo
  # I think Base is not needed at all for this or your example?
  class Base; end

  class One < Base
    include Foo
  end

  class Two < Base
    include Foo
  end
end

foo = Foo::Two.new.as(Foo)

# compiles just fine
case foo
when Foo::One
when Foo::Two
end

module Foo
  class Three
    include Foo # error: Foo is sealed
  end
end
```