@asterite 2020-04-13
> This can lead to hard bugs in case the expressions are not preserved along the way.

This is one of the problems with this PR. However, I prefer the compiler telling me "can't assign ASTNode to Var | InstanceVar | ClassVar", which is kind of easy to understand (ASTNode can be anything), and adding `as(Var | InstanceVar | ClassVar)` at write time, than having to add an `else raise "unreachable"` every time the value is read. This will be specially useful for shards, where values will be constructed by a shard but consumed by others.