@bcardiff 2020-04-14
I think that a module flagged as a `:nodoc:` can be used as a sealed definition. It is not intended to be used, hence removing the scenario of what-if-user-does-X with the module.

With that + handling the exhaustive case for modules I think is fine.

There is always some room for unintended usage of some language construct. Weighting that and the change in semantics depending on how things are written and not what values are involved; I think I prefer the former.

---

NB: In general like the idea of sealing some definitions. Maybe because some familiarity with it from C#. But with the monkey patching it kind of looses the applicability maybe.