@straight-shoota 2020-04-14
I would not derive any semantic meaning from comments. That includes documentation and `:nodoc:`. 
If we want some way to express a sealed type, it should be done with a proper annotation.