@asterite 2020-04-14
> I think that a module flagged as a :nodoc: can be used as a sealed definition

I'm not sure. In the recent example of `Process::Status`, sealed would have come handy to say: I want a type hierarchy, but these are the only possible types (not that we'll merge that, but still an example).

Another example: ASTNode in the compiler. We know all the AST nodes. We use a hierarchy to share common functionality. But the hierarchy is sealed. And ASTNode is far from `:nodoc:`.

I think this is pretty easy to implement: if you use `@[Sealed]` on top of a type, the only definitions can happen inside that same file. That's a bit nicer than in Sorbet because in Sorbet you are forced to put children inside the parent namespace.

Then, if something is `@[Sealed]` we know all the types and can do exhaustive check.

It's a very, very simple change.

But I don't think it's related to this PR. This PR is about nor merging union types and not losing information about an instance var by going to the parent type. The sealed discussion is related to exhaustive case.