@asterite 2020-04-16
@bcardiff Just to be clear, you can't check exhaustiveness if the type is, say, `ASTNode` because anyone can inherit it. This PR makes it so you can have a variable of type `Var | InstanceVar`, something that you couldn't do before. In that case the compiler can check exhaustiveness: it must be either `Var` or `InstanceVar`. Of course these types can be extended, but if `Var` and `InstanceVar` are covered, all subtypes are covered too. This is what the compiler will check.