@HertzDevil 2022-03-31
Here is the discussion that prompted this: https://forum.crystal-lang.org/t/reducible-unions-from-type-filters/4525

There it is mentioned that those unmerged union types may still appear as the result of type filtering on `#is_a?` or `#responds_to?`:

```crystal
class A; end
class B1 < A; end
class B2 < A; end
class B3 < A; end

x = B1.new || B2.new

if x.is_a?(B1) || x.is_a?(B2)
  typeof(x) # => (B1 | B2)
end

if x.is_a?(B1 | B2)
  # since 1.2.0 (#11176)
  typeof(x) # => (B1 | B2)
end
```

This PR, I assume, extends this to arbitrary type expressions (such as for instance variables), and ensures that no variant types inside a union are subtypes of other variant types, whereas `||` still merges unions because no type declarations are possible there. Am I right?

Does the `Union` generic merge?