@asterite 2022-03-31
> Do instances of the Union generic merge?

I can't remember how it behaves in this PR, but if you have `@x : Union(A, B)` then I think it should remain as `A | B` and not go to the parent.

Was that the question?